moved seed data into its own folder
